+++ src/CHANGELOG.md
@@ -43,4 +43,4 @@
 
 * Fix spelling: Rename B9Invokation to B9Invocation
   
-  
+* Rename `FromAST` to `FromAST`